Refactor away the validator package
